Move loader into components directory
